FIX: Update should support set to null.
Now an InvalidOperationException is thrown with message
"Can't use a non-equality here" because the ' == null ' in a
Set is translated to IS NULL. This is correct in a Where statement,
but not in a Set statement.

Added one line of code to SubSonic.Query.Update which checks if a
Comparison.Is has been found. If so, it is changed to Comparison.Equals.

Also added a Unit Test to SubSonic.Tests.BugReports.Update